Added Colllege Basketball, as well as some refactoring
Lots of renaming for consistency with toolStripMenu's as well as
refactoring of code in the MainForm so that it is easier to find methods
in Region blocks as opposed to sporatic.  Still needs improvement in
other classes, but this class is clean now.  Also added College
Basketball as an openable CSV file and team creation.  Additionally
player values can be calculated for Basketball as well.